add blurr line
