+++ frontend/src/views/profiles/index.js
@@ -13,26 +13,9 @@ import {
 import { useTheme } from "@mui/material/styles";
 import { styled } from "@mui/material/styles";
 import ProfileBanner from "./ProfileBanner";
+import ProfileDescription from "./ProfileDescription";
 
 // const about = `I'm the certified coders pimp.\nAs my mentees, you'll be part of the CORE gang (Code-Whore). \nAfter receiving my wisdom, 11/10 of my COREs now work at FAANG companies.`;
-const skills = [
-  "Deep Learning",
-  "ML",
-  "Python",
-  "Go",
-  "Computer Systems",
-  "AI",
-  "Embedded Programming",
-];
-const services = [
-  "Mock Interview",
-  "Code Reviews",
-  "Career Growth",
-  "Break into Tech",
-  "Salary Negotiation",
-  "Promotion",
-  "Team Lead",
-];
 
 const compliment = `Very friendly mentor. Listen to my concern and give good advice. Very realistic, not too overly pessimistic nor optimistic`;
 
@@ -40,26 +23,6 @@ const apiUrl = process.env.REACT_APP_API_URL;
 
 // const Offset = styled('div')(({ theme }) => theme.mixins.toolbar);
 
-const SkillsSection = ({ sectionTitle, list }) => {
-  return (
-    <>
-      <Box>
-        <Typography variant="h5">{sectionTitle}</Typography>
-      </Box>
-      <Stack
-        direction="row"
-        spacing={1}
-        useFlexGap
-        flexWrap="wrap"
-        sx={{ width: "100%" }}
-      >
-        {list.map((item, index) => (
-          <Chip key={index} label={item} />
-        ))}
-      </Stack>
-    </>
-  );
-};
 
 const BackgroundWavesPhoto = ({ children }) => {
   return (
@@ -156,43 +119,7 @@ const ProfileViews = () => {
               avatarUrl={person.picture}
               data={person}
             />
-            <Box
-              sx={{
-                overflow: "scroll",
-                width: "70%",
-                height: "70vh",
-                boxShadow:
-                  "inset 0px -20px 10px -10px rgba(235, 241, 250, 0.95)",
-              }}
-            >
-              {/* Intro */}
-              <Box sx={{ marginBottom: 2 }}>
-                <Typography variant="h5" fontWeight="medium" marginBottom={2}>
-                  About Me
-                </Typography>
-                <Typography>{person.description}</Typography>
-              </Box>
-
-              {/* Ice breakers - hardcoded*/}
-              <Box sx={{ marginBottom: 2 }}>
-                <Typography variant="h6" fontWeight="regular">
-                  Favorite WWU Professor
-                </Typography>
-                <Typography>Yudong</Typography>
-                <Typography variant="h6" fontWeight="regular">
-                  Hardest CS Class
-                </Typography>
-                <Typography>CSCI 247: OS</Typography>
-              </Box>
-
-              {/* Meet me for - hardcoded */}
-              <Box sx={{ marginBottom: 2 }}>
-                <Typography variant="h5" fontWeight="medium" marginBottom={1}>
-                  Meet Me For
-                </Typography>
-                <SkillsSection list={services} />
-              </Box>
-            </Box>
+            <ProfileDescription person={person}/>
           </Stack>
         </Paper>
       </BackgroundWavesPhoto>
